tests/wpt/mozilla/tests/mozilla/window_performance.html

<html>
<head>
  <title></title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
test(function() {
  assert_not_equals(window.performance, undefined);
  assert_true(window.performance instanceof Performance, "Should be Performance");

  assert_not_equals(window.performance.timing, undefined);

  var entries = window.performance.getEntries();
  assert_not_equals(entries.length, 0);
  assert_true(entries[0] instanceof PerformanceNavigationTiming);

  // TODO(#21254) navigationTiming/startTime is not fully implemented yet, so this assertion will fail
  assert_greater_than(entries[0].startTime, 0);

  var last = window.performance.now();
  assert_greater_than(last, 0);

  // Check that window.performance.now() is monotonically increasing
  for (var i = 0; i < 100; i++) {
    var next = window.performance.now();
    assert_greater_than(next, last);
    last = next;
  }
});
</script>
</body>
</html>
